--- bin/test_submission
@@ -1,2 +1,0 @@
-#!/usr/bin/env python
-from rampwf import test_submission  # noqa
