+++ rampwf/__init__.py
@@ -2,4 +2,6 @@ from . import prediction_types
 from . import score_types
 from . import workflows
 
+__all__ = ['kits', 'utils']
+
 __version__ = '0.1.dev'
