+++ rampwf/utils/__init__.py
@@ -1,0 +1,3 @@
+from .testing import assert_submission
+
+__all__ = ['assert_submission']
