+++ setup.py
@@ -6,6 +6,7 @@ import os
 
 # we are using a setuptools namespace
 import setuptools  # noqa
+from setuptools import find_packages
 from numpy.distutils.core import setup
 
 version = None
@@ -35,38 +36,27 @@ if __name__ == "__main__":
     if os.path.exists('MANIFEST'):
         os.remove('MANIFEST')
 
-    setup(
-        name=DISTNAME,
-        maintainer=MAINTAINER,
-        include_package_data=True,
-        maintainer_email=MAINTAINER_EMAIL,
-        description=DESCRIPTION,
-        license=LICENSE,
-        url=URL,
-        version=VERSION,
-        download_url=DOWNLOAD_URL,
-        long_description=open('README.md').read(),
-        zip_safe=False,  # the package can run out of an .egg file
-        classifiers=['Intended Audience :: Science/Research',
-                     'Intended Audience :: Developers',
-                     'License :: OSI Approved',
-                     'Programming Language :: Python',
-                     'Topic :: Software Development',
-                     'Topic :: Scientific/Engineering',
-                     'Operating System :: Microsoft :: Windows',
-                     'Operating System :: POSIX',
-                     'Operating System :: Unix',
-                     'Operating System :: MacOS'],
-        platforms='any',
-        packages=[
-            'rampwf',
-            'rampwf.score_types',
-            'rampwf.tests',
-            'rampwf.prediction_types',
-            'rampwf.workflows',
-        ],
-        scripts=[
-            'bin/test_submission',
-            'bin/test_backend',
-        ]
-    )
+    setup(name=DISTNAME,
+          maintainer=MAINTAINER,
+          include_package_data=True,
+          maintainer_email=MAINTAINER_EMAIL,
+          description=DESCRIPTION,
+          license=LICENSE,
+          url=URL,
+          version=VERSION,
+          download_url=DOWNLOAD_URL,
+          long_description=open('README.md').read(),
+          zip_safe=False,  # the package can run out of an .egg file
+          classifiers=['Intended Audience :: Science/Research',
+                       'Intended Audience :: Developers',
+                       'License :: OSI Approved',
+                       'Programming Language :: Python',
+                       'Topic :: Software Development',
+                       'Topic :: Scientific/Engineering',
+                       'Operating System :: Microsoft :: Windows',
+                       'Operating System :: POSIX',
+                       'Operating System :: Unix',
+                       'Operating System :: MacOS'],
+          platforms='any',
+          packages=find_packages()
+          )
